QuadraticOutputSystems

Collection of tools for working with quadratic output systems defined as

\[\begin{align*} \dot x(t) &= Ax(t) + Bx(t), \\ y(t) &= x(t)^{\mathsf{T}}M x(t), \end{align*}\]

where $A \in \R^{n \times n}$, $B \in \R^{n \times m}$, and $M \in \R^{n \times n}$. As in BenPD2022 we assume $M = M^\mathsf{T}$. This can be imposed without loss of generality.

Functionality

QuadraticOutputSystems.h2norm — Function

h2norm(A, B, M)

Computes the h2norm of quadratic output system $\dot x = Ax+Bu, y = x^{\mathsf{T}} M x$

QuadraticOutputSystems.h2error — Function

h2error(A1, B1, M1, A2, B2, M2)

Computes the h2-error between two quadratic output systems defined by A1, B1, and M1 and A2, B2, and M2, respectively. Note that the optional named argumentd BQ1B, which defaults to B1'*qo_observability_gramian(A1, B1, M1)*B1 can be passed, if BQ1B is known beforehand.

QuadraticOutputSystems.h2error_sqr — Function

h2error_sqr(A1, B1, M1, A2, B2, M2)

Computes the square of the h2-error between two quadratic output systems defined by A1, B1, and M1 and A2, B2, and M2, respectively. Note that the optional named argumentd BQ1B, which defaults to B1'*qo_observability_gramian(A1, B1, M1)*B1 can be passed, if BQ1B is known beforehand.

QuadraticOutputSystems.qo_observability_gramian — Function

qo_observability_gramian(A, B, M)

Computes the quadratic-output observability gramian as defined in BenPD2022 for a quadratic output system with system matrices $A, B, M$.

References

@Article{BenPD2022,
  author={Benner, Peter and Goyal, Pawan and Duff, Igor Pontes},
  journal={IEEE Transactions on Automatic Control}, 
  title={Gramians, Energy Functionals, and Balanced Truncation for Linear Dynamical Systems With Quadratic Outputs}, 
  year={2022},
  volume={67},
  number={2},
  pages={886-893},
  doi={10.1109/TAC.2021.3086319}
}